Adjust type for user id columns

+++ database/migrations/2025_02_25_113647_adjust_team_leader_id_type.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('teams', function (Blueprint $table) {
-            $table->unsignedBigInteger('leader_id')->change();
+            $table->unsignedInteger('leader_id')->change();
         });
     }
